registry refactor
