+++ app/src/registry/mainnet/mainnet.ts
@@ -1,0 +1,326 @@
+import { Registry } from '..'
+import {
+  Acala,
+  AssetHub,
+  Astar,
+  Bifrost,
+  Centrifuge,
+  Ethereum,
+  Hydration,
+  Interlay,
+  Moonbeam,
+  Mythos,
+  Polimec,
+  RelayChain,
+} from './chains'
+import { relayToParaRoutes } from './relayToParaRoutes'
+import { Eth, Polkadot } from './tokens'
+
+/* Mainnet :: Polkadot - Ethereum */
+
+// Assembling the Mainnet registry
+export const REGISTRY: Registry = {
+  chains: [
+    Ethereum,
+    AssetHub,
+    RelayChain,
+    Bifrost,
+    Mythos,
+    Hydration,
+    Acala,
+    Moonbeam,
+    Interlay,
+    Polimec,
+    Centrifuge,
+    Astar,
+  ],
+  tokens: [
+    Eth.WETH,
+    Eth.WBTC,
+    Eth.USDC,
+    Polkadot.USDC,
+    Eth.USDT,
+    Polkadot.USDT,
+    Eth.WSTETH,
+    Eth.TBTC,
+    Polkadot.DOT,
+    Eth.DAI,
+    Polkadot.ACA,
+    Polkadot.CFG,
+    Polkadot.BNC,
+    Polkadot.GLMR,
+    Polkadot.ASTR,
+    Polkadot.INTR,
+    Polkadot.VDOT,
+    Polkadot.IBTC,
+    Eth.MYTH,
+    Polkadot.HDX,
+    Eth.TON,
+    Eth.SHIB,
+    Eth.PEPE,
+  ],
+  routes: [
+    // Snowbridge routes
+    {
+      from: Ethereum.uid,
+      to: AssetHub.uid,
+      sdk: 'SnowbridgeApi',
+      tokens: [
+        Eth.WETH.id,
+        Eth.WBTC.id,
+        Eth.USDC.id,
+        Eth.USDT.id,
+        Eth.DAI.id,
+        Eth.MYTH.id,
+        Eth.WSTETH.id,
+        Eth.TBTC.id,
+        Eth.TON.id,
+        Eth.SHIB.id,
+        Eth.PEPE.id,
+      ],
+    },
+    {
+      from: Ethereum.uid,
+      to: Mythos.uid,
+      sdk: 'SnowbridgeApi',
+      tokens: [Eth.MYTH.id],
+    },
+    {
+      from: Ethereum.uid,
+      to: Moonbeam.uid,
+      sdk: 'SnowbridgeApi',
+      tokens: [Eth.WETH.id, Eth.WBTC.id],
+    },
+    {
+      from: Ethereum.uid,
+      to: Bifrost.uid,
+      sdk: 'SnowbridgeApi',
+      tokens: [Eth.WETH.id],
+    },
+    {
+      from: Ethereum.uid,
+      to: Hydration.uid,
+      sdk: 'SnowbridgeApi',
+      tokens: [Eth.WETH.id, Eth.WBTC.id],
+    },
+    {
+      from: AssetHub.uid,
+      to: Ethereum.uid,
+      sdk: 'SnowbridgeApi',
+      tokens: [
+        Eth.WETH.id,
+        Eth.WBTC.id,
+        Eth.USDC.id,
+        Eth.USDT.id,
+        Eth.DAI.id,
+        Eth.MYTH.id,
+        Eth.WSTETH.id,
+        Eth.TBTC.id,
+        Eth.TON.id,
+        Eth.SHIB.id,
+        Eth.PEPE.id,
+      ],
+    },
+    ...relayToParaRoutes,
+
+    // Para to Relay
+    {
+      from: AssetHub.uid,
+      to: RelayChain.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id],
+    },
+    {
+      from: Acala.uid,
+      to: RelayChain.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id],
+    },
+    {
+      from: Hydration.uid,
+      to: AssetHub.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.USDC.id, Polkadot.USDT.id],
+    },
+    {
+      from: Bifrost.uid,
+      to: RelayChain.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id],
+    },
+    {
+      from: Bifrost.uid,
+      to: AssetHub.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Eth.WETH.id, Polkadot.USDC.id, Polkadot.USDT.id],
+    },
+    {
+      from: Bifrost.uid,
+      to: Centrifuge.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id],
+    },
+    {
+      from: Bifrost.uid,
+      to: Hydration.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.BNC.id, Polkadot.VDOT.id],
+    },
+    {
+      from: Bifrost.uid,
+      to: Moonbeam.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.BNC.id, Polkadot.GLMR.id, Polkadot.VDOT.id],
+    },
+    {
+      from: Bifrost.uid,
+      to: Interlay.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.BNC.id, Polkadot.VDOT.id],
+    },
+    {
+      from: Bifrost.uid,
+      to: Acala.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id],
+    },
+    {
+      from: Interlay.uid,
+      to: RelayChain.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id],
+    },
+
+    // Para to Para
+    {
+      from: AssetHub.uid,
+      to: Bifrost.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Eth.WETH.id, Polkadot.USDC.id, Polkadot.USDT.id],
+    },
+    {
+      from: AssetHub.uid,
+      to: Hydration.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.USDC.id, Polkadot.USDT.id],
+    },
+    {
+      from: AssetHub.uid,
+      to: Moonbeam.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.USDC.id, Polkadot.USDT.id],
+    },
+    {
+      from: Moonbeam.uid,
+      to: Hydration.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.GLMR.id],
+    },
+    {
+      from: Acala.uid,
+      to: Hydration.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.ACA.id, Polkadot.DOT.id],
+    },
+    {
+      from: Hydration.uid,
+      to: Moonbeam.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [
+        Polkadot.DOT.id,
+        Polkadot.HDX.id,
+        Polkadot.GLMR.id,
+        Polkadot.BNC.id,
+        Polkadot.VDOT.id,
+      ],
+    },
+    {
+      from: Hydration.uid,
+      to: Bifrost.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.BNC.id, Polkadot.VDOT.id],
+    },
+    {
+      from: Interlay.uid,
+      to: Hydration.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.VDOT.id, Polkadot.INTR.id, Polkadot.IBTC.id],
+    },
+    {
+      from: Hydration.uid,
+      to: Interlay.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [
+        Polkadot.DOT.id,
+        Polkadot.INTR.id,
+        Polkadot.HDX.id,
+        Polkadot.VDOT.id,
+        Polkadot.IBTC.id,
+      ],
+    },
+    {
+      from: Hydration.uid,
+      to: Acala.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.ACA.id, Polkadot.HDX.id],
+    },
+    {
+      from: Hydration.uid,
+      to: Centrifuge.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.CFG.id],
+    },
+    {
+      from: Hydration.uid,
+      to: Astar.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [
+        Polkadot.DOT.id,
+        Polkadot.ASTR.id,
+        Polkadot.HDX.id,
+        Polkadot.GLMR.id,
+        Polkadot.ACA.id,
+        Polkadot.INTR.id,
+      ],
+    },
+    {
+      from: Astar.uid,
+      to: Hydration.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [
+        Polkadot.DOT.id,
+        Polkadot.ASTR.id,
+        Polkadot.HDX.id,
+        Polkadot.GLMR.id,
+        Polkadot.ACA.id,
+        Polkadot.INTR.id,
+      ],
+    },
+    {
+      from: Astar.uid,
+      to: Moonbeam.uid,
+      sdk: 'ParaSpellApi',
+      tokens: [Polkadot.DOT.id, Polkadot.ASTR.id, Polkadot.GLMR.id],
+    },
+  ],
+  assetUid: new Map([
+    [
+      AssetHub.uid,
+      new Map([
+        [Eth.USDC.id, { symbol: 'USDC.e' }],
+        [Polkadot.USDC.id, { id: '1337' }],
+        [Eth.USDT.id, { symbol: 'USDT.e' }],
+        [Polkadot.USDT.id, { id: '1984' }],
+        [Eth.WETH.id, { symbol: 'WETH.e' }],
+        [Eth.WBTC.id, { symbol: 'WBTC.e' }],
+      ]),
+    ],
+    [
+      Hydration.uid,
+      new Map([
+        [Polkadot.USDC.id, { id: '22' }],
+        [Polkadot.USDT.id, { id: '10' }],
+      ]),
+    ],
+  ]),
+}
